Shift Register shows its internal bits when set to serial load, bug fixes, small optimizations and changes

+++ Logisim-Fork/com/cburch/logisim/std/wiring/Pin.java
@@ -330,8 +330,8 @@ public class Pin extends InstanceFactory {
 			bds = painter.getBounds();
 		int x = bds.getX();
 		int y = bds.getY();
-		int width = bds.getWidth();
-		int height = bds.getHeight();
+		short width = (short) bds.getWidth();
+		short height = (short) bds.getHeight();
 		if (facing == Direction.EAST || facing == Direction.WEST) {
 			width -= 5;
 			if (facing == Direction.WEST)
